[lmi][multimodal] Add stronger typing for chat completions use-cases
